+++ utils/validation/authValidationSchemas.js
@@ -1,0 +1,22 @@
+const Joi = require("joi");
+
+const { emailRegexp } = require("../../models/user");
+
+const registerSchema = Joi.object({
+  name: Joi.string(),
+  email: Joi.string().pattern(emailRegexp).required(),
+  password: Joi.string().min(6).required(),
+  subscription: Joi.string().required(),
+});
+
+const loginSchema = Joi.object({
+  email: Joi.string().pattern(emailRegexp).required(),
+  password: Joi.string().min(6).required(),
+});
+
+const schemas = {
+  registerSchema,
+  loginSchema,
+};
+
+module.exports = { schemas };
